+++ refena/test/unit/async_value_test.dart
@@ -4,13 +4,13 @@ import 'package:test/test.dart';
 void main() {
   group('toString', () {
     test(AsyncData, () {
-      expect(AsyncValue.withData(1).toString(), 'AsyncData<int>(1)');
-      expect(AsyncValue.withData('a').toString(), 'AsyncData<String>(a)');
+      expect(AsyncValue.data(1).toString(), 'AsyncData<int>(1)');
+      expect(AsyncValue.data('a').toString(), 'AsyncData<String>(a)');
     });
 
     test(AsyncError, () {
       expect(
-        AsyncValue<String>.withError('test error', StackTrace.empty).toString(),
+        AsyncValue<String>.error('test error', StackTrace.empty).toString(),
         'AsyncError<String>(test error)',
       );
     });
@@ -23,7 +23,7 @@ void main() {
   group('when', () {
     test('Should return the correct value for data', () {
       expect(
-        AsyncValue.withData(1).when(
+        AsyncValue.data(1).when(
           data: (data) => 'data-$data',
           loading: () => 'loading',
           error: (error, stackTrace) => 'error-$error',
@@ -45,7 +45,7 @@ void main() {
 
     test('Should return the correct value for error', () {
       expect(
-        AsyncValue<String>.withError('test error', StackTrace.empty).when(
+        AsyncValue<String>.error('test error', StackTrace.empty).when(
           data: (data) => 'data-$data',
           loading: () => 'loading',
           error: (error, stackTrace) => 'error-$error',
@@ -67,7 +67,7 @@ void main() {
 
     test('Should use error when data is provided', () {
       expect(
-        AsyncValue<int>.withError('error', StackTrace.empty, 3).when(
+        AsyncValue<int>.error('error', StackTrace.empty, 3).when(
           data: (data) => 'data-$data',
           loading: () => 'loading',
           error: (error, stackTrace) => 'error-$error',
@@ -78,7 +78,7 @@ void main() {
 
     test('Should skip error when data is provided and with skip flag', () {
       expect(
-        AsyncValue<int>.withError('error', StackTrace.empty, 3).when(
+        AsyncValue<int>.error('error', StackTrace.empty, 3).when(
           data: (data) => 'data-$data',
           loading: () => 'loading',
           error: (error, stackTrace) => 'error-$error',